Make library match what's in the template
To start with, I think it would be good to have an MVP which just lets
people add library and remove code from their api tests, see if that's
useful and get feedback before changing anything. Though I have some
questions about how HttpClient trait is meant to be used by teams and it
would be good to see some github search results showing how stuff is
used to check if people will be able to move or if they have customised
stuff a lot.